+++ processor/src/test/fixtures/input/com/example/component/PublicCtorModel.java
@@ -1,0 +1,17 @@
+package com.example.component;
+
+import arez.annotations.Action;
+import arez.annotations.ArezComponent;
+
+@ArezComponent
+public abstract class PublicCtorModel
+{
+  public PublicCtorModel()
+  {
+  }
+
+  @Action
+  public void doStuff( final long time, float someOtherParameter )
+  {
+  }
+}